pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "biopb"
version = "0.1.3"
requires-python = ">=3.8"
authors = [ {name = "Ji Yu", email = "jyu@uchc.edu" } ]
description = "Python bindings of BioPB: open protobuf/gRPC protocols for biomedical image processing"
readme = "README.md"
license = "MIT"
dependencies = [
  "protobuf >= 6.32.1",
  "grpcio",
  "numpy",
]

[project.optional-dependencies]
  test = ["pytest >= 7.0, < 8"]

[project.urls]
Homepage = "https://github.com/biopb/biopb"

[tool.setuptools.packages.find]
where = ["src/python"]
include = ["biopb*"]
